requirements.txt
streamlit
pandas
langchain
langchain-openai
gdown
langchain-experimental
python-dotenv